feat: add letter inserts to plates
Keyboard plates now contain a slot for an 8x8x2mm insert above the OLED
display.

A simple insert can be generated with a single letter on top using the
`letterplate.scad` file.

This can be combined with a colour-change on a 3D printer to create a
tiny readable label to tell keyboard modules apart at a glance.